+++ CHANGELOG.md
@@ -28,6 +28,12 @@ Changelog
 - Updated delegated stake pool status on delegation center screen ([PR 2404](https://github.com/input-output-hk/daedalus/pull/2404))
 - Adjusted hover PopOver styles on the first tile in the delegation center screen ([PR 2386](https://github.com/input-output-hk/daedalus/pull/2386))
 
+## 3.3.2
+
+### Features
+
+- Implemented "Voting registration is over" screen ([PR 2428](https://github.com/input-output-hk/daedalus/pull/2428))
+
 ## 3.3.1
 
 ### Features